rename: types/ 폴더 하위 파일을 도메인 별로 이동
- issues, projects 관련 타입을 해당하는 폴더 안의 types 폴더 안으로 이동
