+++ src/app/(with-sidebar)/issues/[issueId]/summary/_components/ranking/normal-list.tsx
@@ -3,7 +3,7 @@
 import { useEffect, useState } from 'react';
 import Image from 'next/image';
 import { VOTE_TYPE } from '@/constants/issue';
-import type { RankedIdeaDto } from '@/types/report';
+import type { RankedIdeaDto } from '@/app/(with-sidebar)/issues/_types';
 import * as DS from './dialog.styles';
 import * as S from './normal-list.styles';
