+++ src/app/(with-sidebar)/issues/_components/categories/category-card-header.tsx
@@ -10,7 +10,7 @@ import {
   Input,
   Title,
 } from './category-card.styles';
-import { IssueStatus } from '@/types/issue';
+import { IssueStatus } from '@/app/(with-sidebar)/issues/_types';
 import { ISSUE_STATUS } from '@/constants/issue';
 
 interface CategoryCardHeaderProps {
