+++ src/app/(with-sidebar)/issues/_components/idea-card/idea-card.styles.ts
@@ -1,7 +1,7 @@
 import styled from '@emotion/styled';
 import { ISSUE_STATUS, VOTE_TYPE } from '@/constants/issue';
 import { theme } from '@/styles/theme';
-import { IssueStatus } from '@/types/issue';
+import { IssueStatus } from '@/app/(with-sidebar)/issues/_types';
 import { CardStatus } from '../../types/idea';
 
 export const Wrapper = styled.div`
