+++ src/app/(with-sidebar)/issues/_components/idea-card/use-idea-card.ts
@@ -2,7 +2,7 @@ import React, { useCallback, useEffect, useRef, useState } from 'react';
 import type { PointerEventHandler } from 'react';
 import { useTooltipStore } from '@/components/tooltip/use-tooltip-store';
 import { ISSUE_STATUS, VOTE_TYPE } from '@/constants/issue';
-import type { IssueStatus } from '@/types/issue';
+import type { IssueStatus } from '@/app/(with-sidebar)/issues/_types';
 import { useVoteMutation } from '@/hooks/issues';
 import { CardStatus } from '../../types/idea';
