+++ src/app/(with-sidebar)/issues/_types/index.ts
@@ -1,0 +1,4 @@
+export * from './category';
+export * from './idea';
+export * from './issue';
+export * from './report';
