+++ src/app/(with-sidebar)/issues/hooks/use-issue-identity.ts
@@ -1,7 +1,7 @@
 import { useSession } from 'next-auth/react';
 import { DEFAULT_SELF_LABEL } from '@/constants/issue';
 import { getUserIdForIssue } from '@/lib/storage/issue-user-storage';
-import type { IssueMember } from '@/types/issue';
+import type { IssueMember } from '@/app/(with-sidebar)/issues/_types';
 import { useIssueQuery } from '@/hooks/issues';
 
 interface UseIssueIdentityOptions {
