+++ src/app/(with-sidebar)/topics/[topicId]/topic-page-client.tsx
@@ -2,7 +2,7 @@
 
 import { useTopicQuery } from '@/hooks/topics';
 import { useTopicEvents } from '@/hooks/topics/use-topic-events';
-import { IssueConnection, IssueMapData, IssueNode } from '@/types/issue';
+import { IssueConnection, IssueMapData, IssueNode } from '@/app/(with-sidebar)/issues/_types';
 import IssueNodeSkeletonGrid from '../_components/issue-node-skeleton-grid/issue-node-skeleton-grid';
 import TopicCanvas from '../_components/topic-canvas/topic-canvas';
