+++ src/app/(with-sidebar)/topics/_components/issue-node/issue-node.styles.ts
@@ -2,7 +2,7 @@ import styled from '@emotion/styled';
 import { css, keyframes } from '@emotion/react';
 import { ISSUE_STATUS } from '@/constants/issue';
 import { theme } from '@/styles/theme';
-import { IssueStatus } from '@/types/issue';
+import { IssueStatus } from '@/app/(with-sidebar)/issues/_types';
 
 export const NodeContainer = styled.div<{ status: IssueStatus }>`
   display: flex;
