+++ src/app/projects/_components/project-list/project-list.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { ProjectCard } from '@/app/projects/_components/project-card/project-card';
-import type { ProjectListItem } from '@/types/project';
+import type { ProjectListItem } from '@/app/projects/_types';
 import * as S from './project-list.styles';
 
 interface ProjectListProps {
