+++ src/constants/issue.ts
@@ -1,4 +1,4 @@
-import { IssueStatus } from '@/types/issue';
+import { IssueStatus } from '@/app/(with-sidebar)/issues/_types';
 
 export const ISSUE_STATUS = {
   BRAINSTORMING: 'BRAINSTORMING',
