+++ src/hooks/issues/use-category-mutation.ts
@@ -3,7 +3,7 @@ import toast from 'react-hot-toast';
 import { useSseConnectionStore } from '@/app/(with-sidebar)/issues/store/use-sse-connection-store';
 import { CLIENT_ERROR_MESSAGES } from '@/constants/error-messages';
 import { createCategory, deleteCategory, updateCategory } from '@/lib/api/category';
-import type { Category } from '@/types/category';
+import type { Category } from '@/app/(with-sidebar)/issues/_types';
 
 type CategoryPayload = {
   title: string;
