+++ src/hooks/issues/use-category-query.ts
@@ -1,7 +1,7 @@
 import { useQuery } from '@tanstack/react-query';
 import type { Category } from '@/app/(with-sidebar)/issues/types/category';
 import { fetchCategories } from '@/lib/api/category';
-import type { Category as DbCategory } from '@/types/category';
+import type { Category as DbCategory } from '@/app/(with-sidebar)/issues/_types';
 
 export const useCategoryQuery = (issueId: string) => {
   return useQuery({
