+++ src/hooks/topics/use-topic-mutation.ts
@@ -2,7 +2,7 @@ import { useRouter } from 'next/navigation';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 import { toast } from 'react-hot-toast';
 import { type CreateTopicData, createTopic, deleteTopic, updateTopicTitle } from '@/lib/api/topic';
-import type { ProjectwithTopic } from '@/types/project';
+import type { ProjectwithTopic } from '@/app/projects/_types';
 
 export const useCreateTopicMutation = () => {
   const queryClient = useQueryClient();
