+++ src/lib/api/category.ts
@@ -1,4 +1,4 @@
-import type { Category } from '@/types/category';
+import type { Category } from '@/app/(with-sidebar)/issues/_types';
 import getAPIResponseData from '../utils/api-response';
 import { withSseHeader } from '../utils/with-sse-header';
