+++ src/lib/repositories/project.repository.ts
@@ -1,5 +1,5 @@
 import { prisma } from '@/lib/prisma';
-import { PrismaTransaction } from '@/types/prisma';
+import { PrismaTransaction } from '@/types';
 
 type PrismaClientOrTx = PrismaTransaction | typeof prisma;
