+++ src/lib/repositories/report.repository.ts
@@ -1,6 +1,6 @@
 import { prisma } from '@/lib/prisma';
-import { PrismaTransaction } from '@/types/prisma';
-import { ReportWithDetails } from '@/types/report';
+import { PrismaTransaction } from '@/types';
+import { ReportWithDetails } from '@/app/(with-sidebar)/issues/_types';
 
 type PrismaClientOrTx = PrismaTransaction | typeof prisma;
